Blend IMetricMessageSink and IMessageSink
