+++ server/collector/src/main/java/org/bithon/server/collector/sink/kafka/KafkaMetricSink.java
@@ -18,67 +18,82 @@ package org.bithon.server.collector.sink.kafka;
 
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.annotation.OptBoolean;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.ImmutableMap;
+import org.apache.kafka.clients.producer.ProducerConfig;
+import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.common.serialization.StringSerializer;
 import org.bithon.component.commons.utils.CollectionUtils;
-import org.bithon.server.sink.metrics.IMessageSink;
+import org.bithon.component.commons.utils.Preconditions;
 import org.bithon.server.sink.metrics.IMetricMessageSink;
 import org.bithon.server.sink.metrics.SchemaMetricMessage;
-import org.bithon.server.storage.datasource.input.IInputRow;
+import org.springframework.kafka.core.DefaultKafkaProducerFactory;
+import org.springframework.kafka.core.KafkaTemplate;
 
-import java.util.List;
+import java.nio.charset.StandardCharsets;
+import java.util.Map;
 
 /**
+ * Sink message to Kafka.
+ * <p>
+ * The instance is created by {@link org.bithon.server.collector.CollectorAutoConfiguration} and
+ * {@link org.bithon.server.collector.sink.SinkConfig}
+ *
  * @author frank.chen021@outlook.com
  * @date 2021/3/15
  */
 @JsonTypeName("kafka")
 public class KafkaMetricSink implements IMetricMessageSink {
 
-    private final IMessageSink<SchemaMetricMessage> sink;
+    private final KafkaTemplate<String, String> producer;
+    private final ObjectMapper objectMapper;
+    private final String topic;
 
     @JsonCreator
-    public KafkaMetricSink(@JacksonInject(useInput = OptBoolean.FALSE) IMessageSink<SchemaMetricMessage> sink) {
-        this.sink = sink;
+    public KafkaMetricSink(@JsonProperty("props") Map<String, Object> props,
+                           @JacksonInject(useInput = OptBoolean.FALSE) ObjectMapper objectMapper) {
+        this.topic = (String) props.remove("topic");
+        Preconditions.checkNotNull(topic, "topic is not configured for metrics sink");
+
+        this.producer = new KafkaTemplate<>(new DefaultKafkaProducerFactory<>(props,
+                                                                              new StringSerializer(),
+                                                                              new StringSerializer()),
+                                            ImmutableMap.of(ProducerConfig.CLIENT_ID_CONFIG, "metrics"));
+
+        this.objectMapper = objectMapper;
     }
 
     @Override
-    public void process(String messageType, List<IInputRow> messages) {
-        if (CollectionUtils.isEmpty(messages)) {
+    public void process(String messageType, SchemaMetricMessage message) {
+        if (CollectionUtils.isEmpty(message.getMetrics())) {
             return;
         }
-        this.sink.process(messageType, SchemaMetricMessage.builder().metrics(messages).build());
-//
-//        String key = null;
-//
-//        //
-//        // a batch message in written into a single kafka message in which each text line is a single metric message
-//        //
-//        // of course we could also send messages in this batch one by one to Kafka,
-//        // but I don't think it has advantages over the way below
-//        //
-//        StringBuilder messageText = new StringBuilder();
-//        for (IInputRow metricMessage : messages) {
-//            // Sink receives messages from an agent, it's safe to use instance name of first item
-//            key = metricMessage.getColAsString("instanceName");
-//
-//            // deserialization
-//            try {
-//                messageText.append(objectMapper.writeValueAsString(metricMessage));
-//            } catch (JsonProcessingException ignored) {
-//            }
-//
-//            //it's not necessary, only used to improve readability of text when debugging
-//            messageText.append('\n');
-//        }
-//
-//        ProducerRecord<String, String> record = new ProducerRecord<>(this.topic, key, messageText.toString());
-//        record.headers().add("type", messageType.getBytes(StandardCharsets.UTF_8));
-//
-//        this.producer.send(record);
+
+        String appName = message.getMetrics().get(0).getColAsString("appName");
+        String instanceName = message.getMetrics().get(0).getColAsString("instanceName");
+        if (appName == null || instanceName == null) {
+            return;
+        }
+
+        try {
+            String messageText = this.objectMapper.writeValueAsString(message);
+
+            String key = appName + "/" + instanceName;
+            ProducerRecord<String, String> record = new ProducerRecord<>(this.topic, key, messageText);
+            record.headers().add("type", messageType.getBytes(StandardCharsets.UTF_8));
+
+            this.producer.send(record);
+        } catch (JsonProcessingException ignored) {
+        }
     }
 
     @Override
     public void close() {
+        this.producer.destroy();
     }
 }
+
