+++ server/collector/src/main/java/org/bithon/server/collector/source/http/MetricHttpCollector.java
@@ -18,7 +18,7 @@ package org.bithon.server.collector.source.http;
 
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
-import org.bithon.server.sink.metrics.IMessageSink;
+import org.bithon.server.sink.metrics.IMetricMessageSink;
 import org.bithon.server.sink.metrics.MetricMessage;
 import org.bithon.server.sink.metrics.SchemaMetricMessage;
 import org.bithon.server.storage.datasource.DataSourceSchema;
@@ -40,9 +40,9 @@ import java.util.stream.Collectors;
 @ConditionalOnProperty(value = "collector-http.enabled", havingValue = "true")
 public class MetricHttpCollector {
 
-    private final IMessageSink<SchemaMetricMessage> sink;
+    private final IMetricMessageSink sink;
 
-    public MetricHttpCollector(IMessageSink<SchemaMetricMessage> sink) {
+    public MetricHttpCollector(IMetricMessageSink sink) {
         this.sink = sink;
     }
