+++ server/sink/src/main/java/org/bithon/server/sink/metrics/IMetricMessageSink.java
@@ -16,17 +16,17 @@
 
 package org.bithon.server.sink.metrics;
 
+import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import org.bithon.server.storage.datasource.input.IInputRow;
-
-import java.util.List;
 
 /**
  * @author frank.chen021@outlook.com
  * @date 9/12/21 2:17 PM
  */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-
+@JsonSubTypes(value = {
+    @JsonSubTypes.Type(name = "local", value = LocalMetricSink.class),
+})
 public interface IMetricMessageSink extends AutoCloseable {
-    void process(String messageType, List<IInputRow> message);
+    void process(String messageType, SchemaMetricMessage message);
 }
